chore: clean cdes

+++ jpa/src/test/java/com/example/it/JPQLCriteriaBuilderTest.java
@@ -36,6 +36,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
+import java.math.BigDecimal;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -133,9 +134,9 @@ public class JPQLCriteriaBuilderTest {
                             "tuple data :{0},{1},{2},{3},{4},{5},{6},{7}",
                             new Object[]{
                                     result.get(0, String.class),
-                                    result.get(1, Double.class), // hibernate returns BigDecimal
-                                    result.get(2, Double.class), // hibernate returns BigDecimal
-                                    result.get(3, Double.class), // hibernate returns BigDecimal
+                                    result.get(1, BigDecimal.class), // it should return BigDecimal
+                                    result.get(2, BigDecimal.class),
+                                    result.get(3, BigDecimal.class),
                                     result.get(4, Double.class),
                                     result.get(5, Double.class),
                                     result.get(6, Double.class),
